requirements.txt
together
pillow
pydantic==2.10.6
streamlit